Fixed updated data validation parameters
